+++ testcase_new/story/java/src/test/java/com/sequoiadb/lob/basicoperation/TestPutLobByOffset31649.java
@@ -4,8 +4,10 @@ import java.util.Arrays;
 
 import com.sequoiadb.exception.SDBError;
 import com.sequoiadb.lob.utils.RandomWriteLobUtil;
+import com.sequoiadb.testcommon.CommLib;
 import org.bson.types.ObjectId;
 import org.testng.Assert;
+import org.testng.SkipException;
 import org.testng.annotations.AfterClass;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
@@ -36,6 +38,11 @@ public class TestPutLobByOffset31649 extends SdbTestBase {
     @BeforeClass
     public void setUp() {
         sdb = new Sequoiadb( SdbTestBase.coordUrl, "", "" );
+
+        if ( CommLib.isStandAlone( sdb ) ) {
+            throw new SkipException( "is standalone skip testcase" );
+        }
+
         cs = sdb.getCollectionSpace( SdbTestBase.csName );
         if ( cs.isCollectionExist( clName ) ) {
             cs.dropCollection( clName );
